Create README.md
The purpose of this project is to conduct research on how effective genetic algorithms are in board games where typically minimax or alpha beta pruning are applied.  In this project, I have implemented an AI based on genetic algorithms for Checkers.  It turned out to be working admirably. 

Design done by: Redak Stepan